requirements.txt
pandas>=1.5
numpy>=1.21
scikit-learn>=1.2
xgboost>=1.6
joblib>=1.2
imbalanced-learn>=0.10.0
matplotlib>=3.5
streamlit>=1.20.0
google-generativeai>=0.3.2
pyttsx3>=2.90
